rename ignores to nosemgrep, add new semgrepignore test
