+++ semgrep/tests/e2e/snapshots/test_nosemgrep/test_nosem_rule/results.json
