+++ semgrep/tests/e2e/snapshots/test_nosemgrep/test_nosem_rule__invalid_id/error.json
