+++ semgrep/tests/e2e/snapshots/test_nosemgrep/test_nosem_rule__with_disable_nosem/results.json
