+++ semgrep/tests/e2e/snapshots/test_nosemgrep/test_nosem_rule_unicode/results.json
