+++ semgrep/tests/e2e/snapshots/test_nosemgrep/test_regex_rule__nosemgrep/results.json
